Refactor Session Pos
