+++ dayco-app/src/reducers/root.js
@@ -3,6 +3,8 @@ import { combineReducers } from 'redux';
 import { user } from './user';
 import { posts, postsEditModal } from './posts';
 import { alerts } from './alerts';
+import { socket } from './socket';
+
 /**
  * Root reducer for project
  * @module dayco-app/reducers
@@ -11,7 +13,8 @@ const rootReducer = combineReducers({
     user,
     posts,
     postsEditModal,
-    alerts
+    alerts,
+    socket
 });
 
 export default rootReducer;
