+++ dayco-app/src/reducers/socket.js
@@ -1,0 +1,15 @@
+import initialState from '../constants/initialState';
+import * as types from '../constants/types';
+
+
+/**
+ * The posts reducer is responsible
+ * @param {*} state 
+ * @param {*} action 
+ */
+export function socket(state = initialState.socket, action) {
+    switch (action.type) {
+        default:
+            return state;
+    }
+}
